ml_service/requirements.txt
# ML Models Service Requirements

torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
scipy==1.11.4
matplotlib==3.7.2 # Re-added for quantization process
PyYAML==6.0.1
requests==2.31.0
tqdm==4.66.1
thop>=0.1.1
pandas==2.1.3
seaborn==0.13.0 # Re-added for quantization process
setuptools==68.2.2
albumentations==1.3.1
imgaug==0.4.0
flask==3.0.0
gunicorn==21.2.0
openvino-dev
nncf
